Changed structure to make it work with PyPi
