--- cisco_sdwan/__init__.py
@@ -1,1 +1,0 @@
-name = "example_pkg"
