+++ vmanage/command/import_cmd/__init__.py
@@ -1,6 +1,6 @@
 import click
-from .templates import templates
-from .policy import policy
+from vmanage.command.import_cmd.templates import templates
+from vmanage.command.import_cmd.policy import policy
 
 @click.group('import')
 @click.pass_context
