+++ vmanage/command/import_cmd/templates.py
