+++ vmanage/command/show/device.py
