+++ vmanage/command/show/template.py
