--- vmanage_cli/device/__init__.py
@@ -1,12 +1,0 @@
-import click
-from .show import show
-
-@click.group()
-@click.pass_context
-def device(ctx):
-    """
-    Device related commands
-    """
-    print("poop")
-
-device.add_command(show)
